refactor PredictionAdapter
